# goIntermediate/7. corn/docs/swagger.yaml
basePath: /
definitions:
  controller.LoginRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      password:
        example: password1234
        type: string
    required:
    - email
    - password
    type: object
...
  models.User:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  utils.ErrorResponse:
    properties:
      error_msg:
        type: string
      message:
        type: string
    type: object
  utils.LoginResponse:
    properties:
      id:
        type: string
      token:
        type: string
    type: object
  utils.ResponseOK:
    properties:
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://example.com/support
  description: API for managing vouchers
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: Voucher System API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login request payload
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/controller.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResponseOK'
            - properties:
                data:
                  $ref: '#/definitions/utils.LoginResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to save token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Login user
      tags:
      - Authentication
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: User registration request payload
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/utils.ResponseOK'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to register user
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
...
securityDefinitions:
  Authentication:
    in: header
    name: Authorization
    type: apiKey
  UserID:
    in: header
    name: User-ID
    type: apiKey
swagger: "2.0"